+++ poem/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.2.9] 2021-12-22
+
+- Add `Route::try_at`, `Route::try_nest`, `Route::try_nest_no_strip` methods.
+- Add `RouteDomain::try_at` method.
+- Rename `RouteDomain::add` to `RouteDomain::at`.
+
 # [1.2.8] 2021-12-21
 
 - Fix session data is serialized twice. [#109](https://github.com/poem-web/poem/issues/109)